simplify circleci config
