+++ config.json
@@ -1,0 +1,12 @@
+{
+  "name": "nyaaya_cms",
+  "less": "public",
+  "static": "public",
+  "favicon": "public/favicon.ico",
+  "views": "templates/views",
+  "view engine": "pug",
+  "auto update": true,
+  "session": true,
+  "auth": true,
+  "user model": "User"
+}
